Move responses view to its own file for consistency
